Replace bean interface with class and update repository id type

+++ src/main/java/rs/ac/uns/ftn/informatika/repository/ScientificPaperRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
 
 import rs.ac.uns.ftn.informatika.model.ScientificPaper;
 
-public interface ScientificPaperRepository extends ElasticsearchRepository<ScientificPaper, Integer> {
+public interface ScientificPaperRepository extends ElasticsearchRepository<ScientificPaper, Long> {
 }
